requirements.txt
streamlit~=1.31.1
python-dotenv~=1.0.1
langchain~=0.1.6
langchain_openai~=0.0.6
duckduckgo_search~=4.4.3
langchainhub~=0.1.14
arxiv~=2.1.0
g4f~=0.2.1.6